+++ snooplyze/notifier/__init__.py
@@ -1,0 +1,6 @@
+from .notifier import Notifier
+from .console_notifier import ConsoleNotifier
+from .file_notifier import FileNotifier
+
+__all__ = ["Notifier", 
+           "ConsoleNotifier", "FileNotifier"]