+++ snooplyze/notifier/file_notifier.py
@@ -1,0 +1,5 @@
+class FileNotifier:
+
+    def notify(self, message):
+        with open("notifier.log", "a") as f:
+            f.write(f"HIHI: {message}" + "\n")